requirements.txt
backend==0.2.4.1
certifi==2020.12.5
chardet==4.0.0
evdev==1.4.0
idna==2.10
MouseInfo==0.1.3
mss==6.1.0
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.2
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pyTelegramBotAPI==3.7.6
python-xlib==0.29
python3-xlib==0.15
PyTweening==1.0.3
requests==2.25.1
six==1.15.0
urllib3==1.26.3
